ref: Rename centralized_algorithms to cent_algorithms
Rename the centralized_algorithms module to cent_algorithms to align with dst_algorithms, and later fed_algorithms.
